base for hobby module

+++ backend/src/modules/user/user.controller.ts
@@ -1,28 +1,15 @@
 import { Hono } from 'hono';
 import { openApi } from 'hono-zod-openapi';
 import z from 'zod';
-import { DeleteUserSchema, UpdateUserSchema, UserSafeSchema } from './user.dto';
+import {
+  DeleteUserSchema,
+  UpdateUserSchema,
+  UserProfileSchema,
+  UserSafeSchema,
+} from './user.dto';
 import type { AppContext } from '@/src/types';
 
 const userController = new Hono<AppContext>()
-  .get(
-    '/',
-    openApi({
-      tags: ['User', 'Get All'],
-      responses: { 200: z.array(UserSafeSchema) },
-    }),
-    async (c) => {
-      const userService = c.get('services').user;
-      const users = await userService.getAll();
-      return c.json(
-        users.map((user) => ({
-          id: user.id,
-          name: user.name,
-          email: user.email,
-        })),
-      );
-    },
-  )
   .get(
     '/:id',
     openApi({
@@ -33,21 +20,25 @@ const userController = new Hono<AppContext>()
         }),
       },
       responses: {
-        200: UserSafeSchema,
+        200: UserProfileSchema,
         404: z.object({ error: z.string() }),
       },
     }),
     async (c) => {
       const userService = c.get('services').user;
       const id = c.req.param('id');
-      const user = await userService.getById(id);
+      const user = await userService.getProfileById(id);
       if (!user) {
         return c.notFound();
       }
       return c.json({
         id: user.id,
         name: user.name,
-        email: user.email,
+        createdAt: user.createdAt,
+        followedByCount: user._count.followedBy,
+        followsCount: user._count.follows,
+        hobbiesCount: user._count.hobbies,
+        hobbySessionsCount: user._count.hobbySessions,
       });
     },
   )
@@ -59,14 +50,26 @@ const userController = new Hono<AppContext>()
         query: DeleteUserSchema,
       },
       responses: {
-        200: z.object({ deleted: z.boolean() }),
+        204: z.object(),
       },
     }),
     async (c) => {
       const id = c.req.param('id');
+      const currentUserId = c.get('userId');
+      if (currentUserId !== id) {
+        c.status(403);
+        return c.json({ error: 'Unauthorized user' });
+      }
+
       const userService = c.get('services').user;
+
       const deleted = await userService.delete(id);
-      return c.json({ deleted });
+      if (!deleted) {
+        c.status(404);
+        return c.json({ error: 'User not found' });
+      }
+      c.status(204);
+      return c.json(null);
     },
   )
   .patch(
@@ -84,6 +87,13 @@ const userController = new Hono<AppContext>()
     async (c) => {
       const body = c.req.valid('json');
       const id = c.req.param('id');
+
+      const currentUserId = c.get('userId');
+      if (currentUserId !== id) {
+        c.status(403);
+        return c.json({ error: 'Unauthorized user' });
+      }
+
       const userService = c.get('services').user;
       const updatedUser = await userService.update(id, body);
